# docs/swagger.yaml
definitions:
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  api.createAccountRequest:
    properties:
      currency:
        type: string
      owner:
        type: string
    required:
    - currency
    - owner
    type: object
  db.Account:
    properties:
      balance:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      id:
        type: integer
      owner:
        type: string
      updated_at:
        type: string
    type: object
host: nobank.swagger.io
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample bank server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Neobank API
  version: "1.0"
paths:
  /accounts:
    get:
      consumes:
      - application/json
      description: Get a list of accounts with pagination.
      parameters:
      - description: Page ID for pagination
        in: query
        name: page_id
        required: true
        type: integer
      - description: 'Number of accounts per page (min: 5, max: 10)'
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Account'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List accounts
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Create a new account with the specified owner and currency.
      parameters:
      - description: Request body for creating an account
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create an account
      tags:
      - accounts
  /accounts/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve an account by its unique ID.
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Account Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get an account by ID
      tags:
      - accounts
swagger: "2.0"